requirements.txt
oauth2client
google-api-python-client
progress
progressbar2
httplib2shim
google_auth_oauthlib
python-dotenv